% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_surv.R
\name{print.TCGAmutsurv}
\alias{print.TCGAmutsurv}
\title{print for TCGAmutsurv}
\usage{
\method{print}{TCGAmutsurv}(
  x,
  digits = NULL,
  quote = TRUE,
  na.print = NULL,
  print.gap = NULL,
  right = FALSE,
  max = NULL,
  width = NULL,
  useSource = TRUE,
  ...
)
}
\arguments{
\item{x}{instance of TCGAmutsurv}
}
\description{
print for TCGAmutsurv
}
\note{
Individuals may harbor multiple mutations in genes; all events are counted.
}
