% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_surv.R
\name{survfitByGene}
\alias{survfitByGene}
\title{survival fit for TCGAmutsurv}
\usage{
survfitByGene(x, genesym)
}
\arguments{
\item{x}{instance of TCGAmutsurv}
}
\description{
survival fit for TCGAmutsurv
}
\examples{
sk = build_surv_for_mut("SKCM")
sk = survfitByGene(sk, "BRAF")
plot( sk$survfit, lty=1:2, lwd=2, col=c("blue", "orange"))
title("BRAF mutation in SKCM")
}
